At #444 Add rAxis format for RadarChart
